tactile: New and Extended Plots, Methods, and Panel Functions for 'lattice'

Extensions to 'lattice', providing new high-level functions, methods for existing functions, panel functions, and a theme.

See also

Author

Maintainer: Johan Larsson johanlarsson@outlook.com (ORCID)

Other contributors:

Deepayan Sarkar (lattice) [contributor, copyright holder]

Brian Ripley (stats::plot.acf) [contributor]