Panel Function for Ternary Plots

Panel Function for Ternary Plots

Usage

panel.ternaryplot(
  x,
  y,
  z,
  subscripts,
  response = NULL,
  density = FALSE,
  region = density || !is.null(response),
  contour = density || !is.null(response),
  labels = !is.null(response),
  points = TRUE,
  grid = TRUE,
  density_breaks = NULL,
  xlab,
  ylab,
  zlab,
  xlab.default,
  ylab.default,
  zlab.default,
  ...
)

Arguments

x

Numeric vector

y

Numeric vector

z

Numeric vector

subscripts

See lattice::panel.xyplot().

response

An optional response variable

density

Compute two-dimensional density estimates via MASS::kde2d().

region

Fill density or response estimates with a color gradient.

contour

Draw contour lines for density and response estimates.

labels

Label contour lines.

points

Draw points (via panel.ternaryplot.xyplot()).

grid

Draw a reference grid.

density_breaks

Breaks for the density plot if used (see panel.ternaryplot.density()).

xlab

X axis label (the left dimension)

ylab

Y axis label (the right dimension)

zlab

Z axis label (the top dimension)

xlab.default

Internal argument

ylab.default

Internal argument

zlab.default

Internal argument

...

Arguments passed down to subsequent panel functions.

Value

Plots a layer inside a panel of a lattice plot.

See also